+++ resources/views/neura/context-menu/index.blade.php
@@ -3,6 +3,7 @@
 ])
 
 <div x-data="contextMenu" 
+    x-ref="trigger"
     @if(!$disabled) @contextmenu.prevent="open" @endif
     {{ $attributes->merge(['class' => 'inline-block']) }}>
     
@@ -19,7 +20,7 @@
              x-transition:leave="transition ease-in duration-75"
              x-transition:leave-start="opacity-100 scale-100"
              x-transition:leave-end="opacity-0 scale-95"
-             @click.outside="close"
+             @click.stop=""
              @contextmenu.prevent="" 
              @keydown.escape.window="close"
              class="fixed z-50 min-w-[180px] bg-white dark:bg-neutral-800 border border-neutral-200 dark:border-neutral-700 rounded-lg shadow-lg py-1 overflow-hidden"